+++ examples/texteditor/src/main/java/com/example/texteditor/EditorViewModel.kt
@@ -8,7 +8,7 @@ import com.google.gson.Gson
 import dev.yorkie.core.Client
 import dev.yorkie.core.Client.SyncMode.Realtime
 import dev.yorkie.core.Client.SyncMode.RealtimePushOnly
-import dev.yorkie.core.PresenceInfo
+import dev.yorkie.document.presence.PresenceInfo
 import dev.yorkie.document.Document
 import dev.yorkie.document.Document.Event.PresenceChanged
 import dev.yorkie.document.json.JsonText
