+++ yorkie/src/main/kotlin/dev/yorkie/api/ElementConverter.kt
@@ -23,7 +23,7 @@ import dev.yorkie.api.v1.treeNode
 import dev.yorkie.api.v1.treeNodeID
 import dev.yorkie.api.v1.treeNodes
 import dev.yorkie.api.v1.treePos
-import dev.yorkie.core.P
+import dev.yorkie.document.presence.P
 import dev.yorkie.document.crdt.CrdtArray
 import dev.yorkie.document.crdt.CrdtCounter
 import dev.yorkie.document.crdt.CrdtCounter.Companion.asCounterValue
