+++ yorkie/src/main/kotlin/dev/yorkie/document/Document.kt
@@ -3,13 +3,13 @@ package dev.yorkie.document
 import androidx.annotation.VisibleForTesting
 import com.google.protobuf.ByteString
 import dev.yorkie.api.toSnapshot
-import dev.yorkie.core.P
-import dev.yorkie.core.Presence
-import dev.yorkie.core.PresenceChange
-import dev.yorkie.core.PresenceInfo
-import dev.yorkie.core.Presences
-import dev.yorkie.core.Presences.Companion.UninitializedPresences
-import dev.yorkie.core.Presences.Companion.asPresences
+import dev.yorkie.document.presence.P
+import dev.yorkie.document.presence.Presence
+import dev.yorkie.document.presence.PresenceChange
+import dev.yorkie.document.presence.PresenceInfo
+import dev.yorkie.document.presence.Presences
+import dev.yorkie.document.presence.Presences.Companion.UninitializedPresences
+import dev.yorkie.document.presence.Presences.Companion.asPresences
 import dev.yorkie.document.Document.Event.PresenceChanged
 import dev.yorkie.document.Document.Event.PresenceChanged.MyPresence
 import dev.yorkie.document.Document.Event.PresenceChanged.Others
