+++ yorkie/src/main/kotlin/dev/yorkie/document/presence/Presence.kt
@@ -1,4 +1,4 @@
-package dev.yorkie.core
+package dev.yorkie.document.presence
 
 import dev.yorkie.document.change.ChangeContext
